requirements.txt
Django==5.2.7
djangorestframework==3.15.3
django-cors-headers==4.7.0
python-dotenv==1.2.1
Pillow==12.0.0
gunicorn==23.0.0
